pyproject.toml
[project]
name = "searchtolearn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.10.0",
    "google-generativeai>=0.8.4",
    "googletrans>=4.0.2",
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "langchain-google-genai>=2.0.10",
    "pandas>=2.2.3",
    "pypdf>=5.3.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.7",
    "streamlit>=1.42.2",
    "validators>=0.34.0",
]
